Use actor model & add channel list sidebar
