+++ site/styles.css
@@ -1,14 +1,58 @@
 body, html, #app {
   width: 100%;
   height: 100%;
+
   padding: 0;
   margin: 0;
+
+  box-sizing: border-box;
 }
 
 #app {
   display: flex;
-  flex-direction: column;
+  flex-direction: row;
+
   font-family: sans-serif;
+}
+
+#sidebar {
+  flex-grow: 0;
+  width: 240px;
+
+  color: #ccc;
+  background: #222222;
+
+  padding: 18px 0;
+  user-select: none;
+}
+
+#sidebar #channels-list {
+  display: block;
+
+  margin: 0;
+  padding: 0;
+}
+
+#sidebar #channels-list .channel {
+  list-style: none;
+  cursor: pointer;
+
+  padding: 8px 18px;
+}
+
+#sidebar #channels-list .channel.active {
+  color: white;
+}
+
+#sidebar #channels-list .channel:hover {
+  background-color: rgba(255, 255, 255, .2);
+}
+
+#central-pane {
+  flex-grow: 1;
+
+  display: flex;
+  flex-direction: column;
 }
 
 #messages {
